ad-ip-acg/pyproject.toml
[tool.poetry]
name = "ad-ip-acg"
version = "0.1.0"
description = "Python module to take control over external access to your Amazon Workspaces."
authors = ["Thom Rommens"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.36.0"
click = "^8.1.8"
pandas = "^2.2.3"
tabulate = "^0.9.0"
prettyprinter = "^0.18.0"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
